ops/daemon: move `Daemon` struct into `daemon` module
