+++ src/ops/daemon.rs
@@ -1,23 +1,11 @@
 //! Run a BuildLoop for `shell.nix`, watching for input file changes.
 //! Can be used together with `direnv`.
-use crate::build_loop::{self, BuildLoop};
+use crate::daemon::{ping, Daemon};
 use crate::ops::{ok, ExitError, OpResult};
-use crate::project::Project;
-use crate::roots::Roots;
 use crate::socket::communicate::listener;
-use crate::socket::communicate::{CommunicationType, NoMessage, Ping};
-use crate::socket::{ReadError, ReadWriter, Timeout};
-use std::collections::HashMap;
-use std::path::PathBuf;
+use crate::socket::communicate::CommunicationType;
+use crate::socket::ReadWriter;
 use std::sync::mpsc;
-use std::thread;
-
-// TODO: make private again
-/// Instructs the daemon to start a build
-pub struct StartBuild {
-    /// The nix file to watch and build on changes.
-    pub nix_file: PathBuf,
-}
 
 /// See the documentation for lorri::cli::Command::Shell for more
 /// details.
@@ -40,7 +28,7 @@ pub fn main() -> OpResult {
     let (accept_messages_tx, accept_messages_rx) = mpsc::channel();
 
     // TODO join handle
-    let _accept_loop_handle = thread::spawn(move || loop {
+    let _accept_loop_handle = std::thread::spawn(move || loop {
         let accept_messages_tx = accept_messages_tx.clone();
         let _handle = listener
             .accept(|unix_stream, comm_type| match comm_type {
@@ -51,7 +39,7 @@ pub fn main() -> OpResult {
     });
 
     // TODO: join handle
-    let _start_build_loop_handle = thread::spawn(|| {
+    let _start_build_loop_handle = std::thread::spawn(|| {
         for msg in build_messages_rx {
             println!("{:#?}", msg);
         }
@@ -67,76 +55,4 @@ pub fn main() -> OpResult {
 
     // TODO: join all accept handles & accept_loop_handle
     // handle.join().unwrap();
-}
-
-// TODO: move from ops to internals
-/// Keeps all state of the running `lorri daemon` service.
-pub struct Daemon<'a> {
-    // TODO: PathBuf is a nix file
-    /// A thread for each `BuildLoop`, keyed by the nix files listened on.
-    handlers: HashMap<PathBuf, std::thread::JoinHandle<()>>,
-    /// Sending end that we pass to every `BuildLoop` the daemon controls.
-    build_events_tx: mpsc::Sender<build_loop::Event>,
-    /// Static paths the daemon has access to.
-    paths: &'a ::constants::Paths,
-}
-
-impl<'a> Daemon<'a> {
-    /// Create a new daemon. Also return an `mpsc::Receiver` that
-    /// receives `build_loop::Event`s for all builders this daemon
-    /// supervises.
-    pub fn new(paths: &'a ::constants::Paths) -> (Daemon<'a>, mpsc::Receiver<build_loop::Event>) {
-        let (tx, rx) = mpsc::channel();
-        (
-            Daemon {
-                handlers: HashMap::new(),
-                build_events_tx: tx,
-                paths,
-            },
-            rx,
-        )
-    }
-
-    /// Add nix file to the set of files this daemon watches
-    /// & builds if they change.
-    pub fn add(&mut self, nix_file: PathBuf) {
-        let tx = self.build_events_tx.clone();
-        let root_dir = self.paths.gc_root_dir().to_owned();
-
-        self.handlers.entry(nix_file.clone()).or_insert_with(|| {
-            // TODO: refactor Project/Roots stuff, a little bit too complicated
-            // TODO: all these clones are not needed
-            let project = Project::load(nix_file.clone(), root_dir).unwrap();
-            // TODO
-            let roots = Roots::from_project(&project).unwrap();
-            let mut build_loop = BuildLoop::new(nix_file.clone(), roots);
-
-            thread::spawn(move || {
-                // cloning the tx means the daemon’s rx gets all
-                // messages from all builders.
-                build_loop.forever(tx);
-            })
-        });
-    }
-}
-
-/// Accept handler for `socket::communicate::Ping` messages.
-/// For a valid ping message, it sends an instruction to start
-/// the build to `build_chan`.
-// TODO: make private again
-// the ReadWriter here has to be the inverse of the `Client.ping()`, which is `ReadWriter<!, Ping>`
-pub fn ping(rw: ReadWriter<Ping, NoMessage>, build_chan: mpsc::Sender<StartBuild>) {
-    // TODO: read timeout
-    let ping: Result<Ping, ReadError> = rw.read(&Timeout::Infinite);
-    match ping {
-        Err(e) => debug!("didn’t receive a ping!! {:?}", e),
-        Ok(p) => {
-            eprintln!("pinged with {}", p.nix_file.display());
-            build_chan
-                .send(StartBuild {
-                    nix_file: p.nix_file,
-                })
-                .expect("StartBuild channel closed")
-        }
-    }
 }
